CountryRole component moved to its own file
