+++ client/components/EditUserForm/CountryRoles/CountryRole/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CountryRole'
