+++ client/components/EditUserForm/CountryRoles/CountryRoles.tsx
@@ -1,5 +1,5 @@
 import './CountryRoles.scss'
-import React, { MouseEventHandler, useCallback, useMemo, useState } from 'react'
+import React, { useCallback, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { CountryIso, Region, RegionCode } from '@meta/area'
@@ -9,6 +9,8 @@ import { useCountries } from '@client/store/assessment'
 import { useAssessment, useCycle, useSecondaryRegion } from '@client/store/assessment/hooks'
 import { useUser } from '@client/store/user'
 import CountrySelectModal from '@client/components/CountrySelectModal'
+
+import CountryRole from './CountryRole'
 
 // properties used to render ui form fields
 const roles = [
@@ -23,40 +25,6 @@ type ModalOptionsProps = {
   initialSelection: Array<string>
   unselectableCountries: Array<string>
   role: RoleName | null
-}
-
-type CountryRoleProps = {
-  onClick: MouseEventHandler<HTMLButtonElement>
-  role: RoleName
-  user: User
-}
-
-const CountryRole: React.FC<CountryRoleProps> = ({ onClick, role, user }) => {
-  const { i18n } = useTranslation()
-  const userInfo = useUser()
-
-  return (
-    <div className="edit-user__form-field-role-container validation-error-sensitive-field">
-      <div className="edit-user__form-field-role">
-        <div className="role">{i18n.t<string>(Users.getI18nRoleLabelKey(role))}</div>
-        {Users.isAdministrator(userInfo) && (
-          <button className="btn-xs btn-primary" onClick={onClick} type="button">
-            {i18n.t<string>('description.edit')}
-          </button>
-        )}
-      </div>
-
-      <div className="edit-user__form-field-role edit-user__form-field-role-countries">
-        {(user.roles || [])
-          .filter((userRole: UserRole<RoleName>) => userRole.role === role)
-          .map((userRole: UserRole<RoleName>) => (
-            <div key={userRole.countryIso} className="edit-user__form-field-country-box">
-              {i18n.t<string>(`area.${userRole.countryIso}.listName`)}
-            </div>
-          ))}
-      </div>
-    </div>
-  )
 }
 
 type Props = {
